# feat: Add configurable FPS and JPEG quality optimization