--- a/src/main/resources/application.properties
+++ b/src/main/resources/application.properties
@@ -48,3 +48,22 @@ logging.pattern.file=%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{50} 
 # Additional logging for troubleshooting
 logging.level.org.bytedeco.javacv=WARN
 logging.level.org.springframework.web=INFO
+
+# Screen Capture Performance Configuration
+# FPS Settings (5-30 range for performance optimization)
+screenai.capture.fps=15
+screenai.capture.fps.min=5
+screenai.capture.fps.max=30
+
+# JPEG Quality Settings (30-100 range, higher = better quality but larger files)
+screenai.capture.jpeg.quality=70
+screenai.capture.jpeg.quality.min=30
+screenai.capture.jpeg.quality.max=100
+
+# Performance optimization options
+screenai.capture.optimize.ultrafast=true
+screenai.capture.optimize.zerolatency=true
+
+# Circuit breaker configuration for stability
+screenai.capture.circuit-breaker.failure-threshold=10
+screenai.capture.circuit-breaker.timeout=5000