- Increase frame rate from 10 to 15 FPS (50% improvement)
- Add configurable JPEG quality control (70% default)
- Implement FFmpeg ultrafast and zerolatency optimization settings
- Add performance monitoring via /api/status endpoint
- Replace hard-coded values with @Value annotations for dynamic configuration

Performance improvements:
- Reduced frame interval from 100ms to 66ms
- Configurable compression quality for bandwidth optimization
- Enhanced API monitoring for real-time performance tracking

Configuration parameters:
- screenai.capture.frame-rate: 5-30 FPS range
- screenai.capture.jpeg-quality: 30-100% range
- screenai.capture.ultra-fast-mode: FFmpeg preset optimization
- screenai.capture.zero-latency-mode: Low-latency streaming